+++ BinarySearchTree.h
@@ -4,24 +4,24 @@
 
 typedef struct WinnersTree{
     char *nameOfTeam;
-    float points;
+    double points;
     struct WinnersTree *left, *right;
 }WinnersTree;
 
 typedef struct AVLNodeBuilding{
     char *nameOfTeam;
-    float points;
+    double points;
     struct AVLNodeBuilding *left, *right;
     int nodeHeight;
 }AVLNode;
 
-WinnersTree *createTheNode(char *nameOfTeam, float points);
-WinnersTree *addTeamInTree(WinnersTree *BinarySearchTreeRoot, char *nameOfTeam, float points);
+WinnersTree *createTheNode(char *nameOfTeam, double points);
+WinnersTree *addTeamInTree(WinnersTree *BinarySearchTreeRoot, char *nameOfTeam, double points);
 void inorder(WinnersTree *BinarySearchTreeRoot, FILE **outputFile);
 void deleteTheTree(WinnersTree **BinarySearchTreeRoot);
-AVLNode *createAVLNode(char *nameOfTeam, float points);
+AVLNode *createAVLNode(char *nameOfTeam, double points);
 int getTheHeight(AVLNode *AVLRoot);
 AVLNode *rotateToRight(AVLNode *rootSubtree);
 AVLNode *rotateToLeft(AVLNode *rootSubtree);
-AVLNode *insertInAVL(AVLNode *AVLRoot, char *nameOfTeam, float points);
+AVLNode *insertInAVL(AVLNode *AVLRoot, char *nameOfTeam, double points);
 int max(int x, int y);
